Extract out combineRedundantPatches
