+++ src/chronicle/components/EditGroups.tsx
@@ -6,8 +6,7 @@ import { TinyEssayEditor } from "@/tee/components/TinyEssayEditor";
 import * as A from "@automerge/automerge/next";
 import { Hash } from "./Hash";
 import { diffWithProvenance, useActorIdToAuthorMap } from "../utils";
-import { sortBy } from "lodash";
-import { debug } from "console";
+import { combineRedundantPatches } from "../utils";
 
 const inferDiffBase = (doc: A.Doc<MarkdownDoc>) => {
   const changes = A.getAllChanges(doc);
@@ -45,7 +44,7 @@ export const EditGroupsPlayground: React.FC<{ docUrl: AutomergeUrl }> = ({
 
     return {
       ...diff,
-      patches: filterUndoPatches(diff.patches)
+      patches: combineRedundantPatches(diff.patches)
     }
   }, [doc, diffBase]);
 
@@ -83,88 +82,5 @@ export const EditGroupsPlayground: React.FC<{ docUrl: AutomergeUrl }> = ({
   );
 };
 
-const filterUndoPatches = (patches: A.Patch[]) => {
-  let filteredPatches: A.Patch[] = [];
-
-  for (let i = 0; i < patches.length; i++) {
-    let currentPatch = patches[i];
-    let nextPatch = patches[i + 1];
-
-    // Skip if the current and next patches cancel each other out
-    if (
-      nextPatch &&
-      currentPatch.path[0] === "content" &&
-      nextPatch.path[0] === "content" &&
-      currentPatch.action === 'splice' &&
-      nextPatch.action === 'del' &&
-      currentPatch.path[1] === ((nextPatch.path[1] as number) - currentPatch.value.length)
-    ) {
-      const deleted = nextPatch.removed
-      const inserted = currentPatch.value
-
-      // if they have different length we have to be careful and see if they match at the start or end
-      if (inserted.length > deleted.length) {
-        if (inserted.startsWith(deleted)) {
-          const partialInsertPatch = {
-            ...currentPatch,
-            path: ["content", currentPatch.path[1] + deleted.length],
-            value: inserted.slice(deleted.length)
-          }
-          filteredPatches.push(partialInsertPatch)
-
-          i++; // Skip patches
-          continue;      
-        } else if (inserted.endsWith(deleted)) {
-          const partialInsertPatch = {
-            ...currentPatch,
-            value: inserted.slice(0, inserted.length - deleted.length)
-          }
-          filteredPatches.push(partialInsertPatch)
-
-          i++; // Skip patches
-          continue;
-        } 
-
-      } else if (deleted.length > inserted.length) {
-        if (deleted.startsWith(inserted)) {  
-          const removed = deleted.slice(inserted.length)
-          const partialDeletePatch = {
-            ...nextPatch,
-            length: removed.length,
-            removed
-          }
-          filteredPatches.push(partialDeletePatch)
-
-          i++; // Skip patches
-          continue;
-        } else if (deleted.endsWith(inserted)) {
-          const removed = deleted.slice(0, deleted.length - inserted.length)
-          const partialDeletePatch = {
-            ...nextPatch,
-            length: removed.length,
-            removed
-          }
-          filteredPatches.push(partialDeletePatch)
-
-          i++; // Skip patches
-          continue;
-        } 
-      } else if (deleted === inserted) {
-        i++; // Skip patches
-        continue;
-      }
-
-    }
-
-    // If the patches don't cancel each other out, add the current patch to the filtered patches
-    filteredPatches.push(currentPatch);
-  }
-
-  const result = sortBy(filteredPatches, (patch) => patch.path[1]);
-
-  console.log(patches, result)
-
-  return result
-};
