+++ src/chronicle/utils.ts
@@ -4,6 +4,7 @@ import * as A from "@automerge/automerge/next";
 import { useEffect, useRef } from "react";
 import { useForceUpdate } from "@/lib/utils";
 import { useHandle } from "@automerge/automerge-repo-react-hooks";
+import { sortBy } from "lodash";
 
 // Turns hashes (eg for changes and actors) into colors for scannability
 export const hashToColor = (hash: string) => {
@@ -88,4 +89,85 @@ export const useActorIdToAuthorMap = (
   }, []);
 
   return actorIdToAuthorRef.current;
+};
+
+// eliminates redudant patches like a insert followed by and delete of the same characters
+export const combineRedundantPatches = (patches: A.Patch[]) => {
+  let filteredPatches: A.Patch[] = [];
+
+  for (let i = 0; i < patches.length; i++) {
+    let currentPatch = patches[i];
+    let nextPatch = patches[i + 1];
+
+    // Skip if the current and next patches cancel each other out
+    if (
+      nextPatch &&
+      currentPatch.path[0] === "content" &&
+      nextPatch.path[0] === "content" &&
+      currentPatch.action === 'splice' &&
+      nextPatch.action === 'del' &&
+      currentPatch.path[1] === ((nextPatch.path[1] as number) - currentPatch.value.length)
+    ) {
+      const deleted = nextPatch.removed
+      const inserted = currentPatch.value
+
+      // if they have different length we have to be careful and see if they match at the start or end
+      if (inserted.length > deleted.length) {
+        if (inserted.startsWith(deleted)) {
+          const partialInsertPatch = {
+            ...currentPatch,
+            path: ["content", currentPatch.path[1] + deleted.length],
+            value: inserted.slice(deleted.length)
+          }
+          filteredPatches.push(partialInsertPatch)
+
+          i++; // Skip patches
+          continue;      
+        } else if (inserted.endsWith(deleted)) {
+          const partialInsertPatch = {
+            ...currentPatch,
+            value: inserted.slice(0, inserted.length - deleted.length)
+          }
+          filteredPatches.push(partialInsertPatch)
+
+          i++; // Skip patches
+          continue;
+        } 
+
+      } else if (deleted.length > inserted.length) {
+        if (deleted.startsWith(inserted)) {  
+          const removed = deleted.slice(inserted.length)
+          const partialDeletePatch = {
+            ...nextPatch,
+            length: removed.length,
+            removed
+          }
+          filteredPatches.push(partialDeletePatch)
+
+          i++; // Skip patches
+          continue;
+        } else if (deleted.endsWith(inserted)) {
+          const removed = deleted.slice(0, deleted.length - inserted.length)
+          const partialDeletePatch = {
+            ...nextPatch,
+            length: removed.length,
+            removed
+          }
+          filteredPatches.push(partialDeletePatch)
+
+          i++; // Skip patches
+          continue;
+        } 
+      } else if (deleted === inserted) {
+        i++; // Skip patches
+        continue;
+      }
+
+    }
+
+    // If the patches don't cancel each other out, add the current patch to the filtered patches
+    filteredPatches.push(currentPatch);
+  }
+
+  return sortBy(filteredPatches, (patch) => patch.path[1]);
 };
